# This is a combination of 3 commits. # This is the 1st commit message:
MTX-344

* Add monitorType and source as attributes in MappingProcessor

# This is the commit message #2:

MTX-344

* Add JobInfo class

* Add missing javaDoc

* Refactor code after review

# This is the commit message #3:

MTX-344

* Developed sources and computes processing

+++ matrix-engine/src/main/java/com/sentrysoftware/matrix/strategy/utils/RetryOperation.java
@@ -20,7 +20,7 @@ public class RetryOperation<T> {
 	 * Time to wait in seconds before triggering a retry
 	 */
 	@Builder.Default
-	private int waitStrategy = 0;
+	private long waitStrategy = 0;
 
 	/**
 	 * Default value to return if the retry fails
